+++ CHANGELOG.md
@@ -13,6 +13,7 @@ You can also sponsor Tamer here https://buymeacoffee.com/6sjde6vkzl for his inva
 - [matter]: Add list of device types and cluster servers used in the plugin for the certification.
 - [plugin]: Add the possibility to validate shelly wifi devices by name, id and mac (i.e. you can put in the blacklist and whitelist the device name given in the web UI, the shelly id or the shelly mac).
 - [plugin]: Add the possibility to validate shelly BLU devices by name and mac (i.e. you can put in the blacklist and whitelist the device name given in the gateway web UI or the shelly BLU mac). If you put a shelly BLU in the white list, you need to put also his BLU gateway on the white list.
+- [plugin]: Refactor child endpoint creation in ShellyPlatform to use addChildDeviceType and addRequiredClusterServers.
 
 ### Changed